requirements.txt
# Core Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
biopython>=1.79

# Machine Learning and AI
scikit-learn>=0.24.2
tensorflow>=2.6.0
torch>=1.9.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Blockchain Integration
web3>=5.24.0
eth-account>=0.5.6
ipfshttpclient>=0.8.0

# API and Web
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0

# Testing and Development
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
mypy>=0.910

# Documentation
sphinx>=4.1.2
sphinx-rtd-theme>=0.5.2

# Jupyter
jupyter>=1.0.0
notebook>=6.4.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.62.0
pydantic>=1.8.2 